requirements.txt
Flask==2.2.5
PyJWT
bcrypt
requests
